changed serial type to Long

+++ app/src/main/java/com/example/StressOverflow/Item/AddItemFragment.java
@@ -242,7 +242,7 @@ public class AddItemFragment extends DialogFragment{
                                     comments,
                                     newTags,
                                     emptyPictureURLs,
-                                    Integer.valueOf(serial),
+                                    Long.valueOf(serial),
                                     owner
                             ));
                         } catch (IllegalArgumentException e) {
